+++ console2/src/components/molecules/GlobalNavMenu/index.tsx
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -26,12 +26,14 @@ export type GlobalNavTab = 'process' | 'org' | null;
 interface Props {
     activeTab: GlobalNavTab;
     userDisplayName?: string;
+    openDocumentation: () => void;
+    openAbout: () => void;
     logOut: () => void;
 }
 
 class GlobalNavMenu extends React.PureComponent<Props> {
     render() {
-        const { activeTab, userDisplayName, logOut } = this.props;
+        const { activeTab, userDisplayName, openDocumentation, openAbout, logOut } = this.props;
 
         return (
             <Menu fluid={true} inverted={true} size="small" secondary={true}>
@@ -44,11 +46,33 @@ class GlobalNavMenu extends React.PureComponent<Props> {
                 <Menu.Item active={activeTab === 'org'}>
                     <Link to="/org">Organizations</Link>
                 </Menu.Item>
-                <Menu.Item position="right" as={Dropdown} text={userDisplayName}>
-                    {/* TODO can't add an icon here */}
-                    <Dropdown.Menu>
-                        <Dropdown.Item icon="log out" text="Log out" onClick={() => logOut()} />
-                    </Dropdown.Menu>
+                <Menu.Item position="right" fitted="vertically">
+                    <Menu inverted={true} size="small" secondary={true}>
+                        <Menu.Item as={Dropdown} text="System">
+                            <Dropdown.Menu>
+                                <Dropdown.Item
+                                    icon="help"
+                                    text="Documentation"
+                                    onClick={() => openDocumentation()}
+                                />
+                                <Dropdown.Item
+                                    icon="info"
+                                    text="About"
+                                    onClick={() => openAbout()}
+                                />
+                            </Dropdown.Menu>
+                        </Menu.Item>
+                        <Menu.Item as={Dropdown} text={userDisplayName}>
+                            {/* TODO can't add an icon here */}
+                            <Dropdown.Menu>
+                                <Dropdown.Item
+                                    icon="log out"
+                                    text="Log out"
+                                    onClick={() => logOut()}
+                                />
+                            </Dropdown.Menu>
+                        </Menu.Item>
+                    </Menu>
                 </Menu.Item>
             </Menu>
         );