+++ console2/src/components/pages/AboutPage/index.tsx
@@ -18,23 +18,27 @@
  * =====
  */
 import * as React from 'react';
-import { storiesOf } from '@storybook/react';
-import { action } from '@storybook/addon-actions';
-// import StoryRouter from 'storybook-react-router';
+import { Breadcrumb } from 'semantic-ui-react';
 
-import GlobalNavMenu from './';
+import { BreadcrumbSegment } from '../../molecules';
+import { ServerVersion } from '../../organisms';
 
-storiesOf('molecules/GlobalNavMenu', module)
-    // .addDecorator(StoryRouter())
-    // TODO: Wrap Component in a Router context
-    .add('Default', () => (
-        <>
-            <GlobalNavMenu
-                activeTab="process"
-                userDisplayName={'Test Dummy'}
-                openDocumentation={() => action('Clicked Docs')}
-                openAbout={() => action('Clicked About')}
-                logOut={() => action('Clicked Logout')}
-            />
-        </>
-    ));
+class ProcessPage extends React.PureComponent {
+    render() {
+        return (
+            <>
+                <BreadcrumbSegment>
+                    <BreadcrumbSegment>
+                        <Breadcrumb.Section active={true}>About</Breadcrumb.Section>
+                    </BreadcrumbSegment>
+                </BreadcrumbSegment>
+
+                <p>
+                    Server version: <ServerVersion />
+                </p>
+            </>
+        );
+    }
+}
+
+export default ProcessPage;